@@ -10,12 +10,12 @@
 #include <fstream>
 #include <streambuf>
 #include <algorithm>
-#include <boost/regex.hpp>
+#include <regex>
 
 #include <boost/algorithm/string.hpp>
 #include <boost/algorithm/string/join.hpp>
 
-#include <simgrid/msg.h>
+#include <simgrid/s4u.hpp>
 
 #include <rapidjson/document.h>
 #include <rapidjson/writer.h>
@@ -30,14 +30,11 @@ XBT_LOG_NEW_DEFAULT_CATEGORY(jobs, "jobs"); //!< Logging
 
 JobIdentifier::JobIdentifier(const std::string & workload_name,
                              const std::string & job_name) :
-    workload_name(workload_name),
-    job_name(job_name)
+    _workload_name(workload_name),
+    _job_name(job_name)
 {
-    XBT_DEBUG("Parsed workload_name: '%s'", this->workload_name.c_str());
-    XBT_DEBUG("Parsed job_name: '%s'", this->job_name.c_str());
-    XBT_DEBUG("Parsed job_identifier: '%s'", this->to_string().c_str());
-
     check_lexically_valid();
+    _representation = representation();
 }
 
 JobIdentifier::JobIdentifier(const std::string & job_id_str)
@@ -52,18 +49,21 @@ JobIdentifier::JobIdentifier(const std::string & job_id_str)
                "parts, the second one being any string without '!'. Example: 'some_text!42'.",
                job_id_str.c_str());
 
-    this->workload_name = job_identifier_parts[0];
-    XBT_DEBUG("Parsed workload_name: '%s'", this->workload_name.c_str());
-    this->job_name = job_identifier_parts[1];
-    XBT_DEBUG("Parsed job_name: '%s'", this->job_name.c_str());
-    XBT_DEBUG("Parsed job_identifier: '%s'", this->to_string().c_str());
+    this->_workload_name = job_identifier_parts[0];
+    this->_job_name = job_identifier_parts[1];
 
     check_lexically_valid();
+    _representation = representation();
 }
 
 std::string JobIdentifier::to_string() const
 {
-    return workload_name + '!' + job_name;
+    return _representation;
+}
+
+const char *JobIdentifier::to_cstring() const
+{
+    return _representation.c_str();
 }
 
 bool JobIdentifier::is_lexically_valid(std::string & reason) const
@@ -71,16 +71,16 @@ bool JobIdentifier::is_lexically_valid(std::string & reason) const
     bool ret = true;
     reason.clear();
 
-    if(workload_name.find('!') != std::string::npos)
+    if(_workload_name.find('!') != std::string::npos)
     {
         ret = false;
-        reason += "Invalid workload_name '" + workload_name + "': contains a '!'.";
+        reason += "Invalid workload_name '" + _workload_name + "': contains a '!'.";
     }
 
-    if(job_name.find('!') != std::string::npos)
+    if(_job_name.find('!') != std::string::npos)
     {
         ret = false;
-        reason += "Invalid job_name '" + job_name + "': contains a '!'.";
+        reason += "Invalid job_name '" + _job_name + "': contains a '!'.";
     }
 
     return ret;
@@ -92,14 +92,38 @@ void JobIdentifier::check_lexically_valid() const
     xbt_assert(is_lexically_valid(reason), "%s", reason.c_str());
 }
 
+string JobIdentifier::workload_name() const
+{
+    return _workload_name;
+}
+
+string JobIdentifier::job_name() const
+{
+    return _job_name;
+}
+
+string JobIdentifier::representation() const
+{
+    return _workload_name + '!' + _job_name;
+}
+
 bool operator<(const JobIdentifier &ji1, const JobIdentifier &ji2)
 {
     return ji1.to_string() < ji2.to_string();
 }
 
+bool operator==(const JobIdentifier &ji1, const JobIdentifier &ji2)
+{
+    return ji1.to_string() == ji2.to_string();
+}
+
+std::size_t JobIdentifierHasher::operator()(const JobIdentifier & id) const
+{
+    return std::hash<std::string>()(id.to_string());
+}
 
 
-BatTask::BatTask(Job * parent_job, Profile * profile) :
+BatTask::BatTask(JobPtr parent_job, ProfilePtr profile) :
     parent_job(parent_job),
     profile(profile)
 {
@@ -122,9 +146,9 @@ void BatTask::compute_leaf_progress()
     {
         if (ptask != nullptr) // The parallel task has already started
         {
-            // WARNING: This is not returning the flops amount but the remainin quantity of work
+            // WARNING: 'get_remaining_ratio' is not returning the flops amount but the remaining quantity of work
             // from 1 (not started yet) to 0 (completely finished)
-            current_task_progress_ratio = 1 - MSG_task_get_remaining_work_ratio(ptask);
+            current_task_progress_ratio = 1 - ptask->get_remaining_ratio();
         }
         else
         {
@@ -137,7 +161,7 @@ void BatTask::compute_leaf_progress()
 
         double runtime = simgrid::s4u::Engine::get_clock() - delay_task_start;
 
-        // Manages empty delay job (why?!)
+        // Manages empty delay job (TODO why?!)
         if (delay_task_required == 0)
         {
             current_task_progress_ratio = 1;
@@ -178,10 +202,7 @@ BatTask* Job::compute_job_progress()
 
 Jobs::~Jobs()
 {
-    for (auto mit : _jobs)
-    {
-        delete mit.second;
-    }
+    _jobs.clear();
 }
 
 void Jobs::set_profiles(Profiles *profiles)
@@ -207,53 +228,69 @@ void Jobs::load_from_json(const rapidjson::Document &doc, const std::string &fil
     {
         const Value & job_json_description = jobs[i];
 
-        Job * j = Job::from_json(job_json_description, _workload, error_prefix);
+        auto j = Job::from_json(job_json_description, _workload, error_prefix);
 
         xbt_assert(!exists(j->id), "%s: duplication of job id '%s'",
                    error_prefix.c_str(), j->id.to_string().c_str());
         _jobs[j->id] = j;
+        _jobs_met.insert({j->id, true});
     }
 }
 
-Job *Jobs::operator[](JobIdentifier job_id)
+JobPtr Jobs::operator[](JobIdentifier job_id)
 {
     auto it = _jobs.find(job_id);
     xbt_assert(it != _jobs.end(), "Cannot get job '%s': it does not exist",
-               job_id.to_string().c_str());
+               job_id.to_cstring());
     return it->second;
 }
 
-const Job *Jobs::operator[](JobIdentifier job_id) const
+const JobPtr Jobs::operator[](JobIdentifier job_id) const
 {
     auto it = _jobs.find(job_id);
     xbt_assert(it != _jobs.end(), "Cannot get job '%s': it does not exist",
-               job_id.to_string().c_str());
+               job_id.to_cstring());
     return it->second;
 }
 
-Job *Jobs::at(JobIdentifier job_id)
+JobPtr Jobs::at(JobIdentifier job_id)
 {
     return operator[](job_id);
 }
 
-const Job *Jobs::at(JobIdentifier job_id) const
+const JobPtr Jobs::at(JobIdentifier job_id) const
 {
     return operator[](job_id);
 }
 
-void Jobs::add_job(Job *job)
+void Jobs::add_job(JobPtr job)
 {
     xbt_assert(!exists(job->id),
                "Bad Jobs::add_job call: A job with name='%s' already exists.",
-               job->id.to_string().c_str());
+               job->id.to_cstring());
 
     _jobs[job->id] = job;
+    _jobs_met.insert({job->id, true});
 }
 
-bool Jobs::exists(JobIdentifier job_id) const
+void Jobs::delete_job(const JobIdentifier & job_id, const bool & garbage_collect_profiles)
 {
-    auto it = _jobs.find(job_id);
-    return it != _jobs.end();
+    xbt_assert(exists(job_id),
+               "Bad Jobs::delete_job call: The job with name='%s' does not exist.",
+               job_id.to_cstring());
+
+    std::string profile_name = _jobs[job_id]->profile->name;
+    _jobs.erase(job_id);
+    if (garbage_collect_profiles)
+    {
+        _workload->profiles->remove_profile(profile_name);
+    }
+}
+
+bool Jobs::exists(const JobIdentifier & job_id) const
+{
+    auto it = _jobs_met.find(job_id);
+    return it != _jobs_met.end();
 }
 
 bool Jobs::contains_smpi_job() const
@@ -261,8 +298,8 @@ bool Jobs::contains_smpi_job() const
     xbt_assert(_profiles != nullptr, "Invalid Jobs::containsSMPIJob call: setProfiles had not been called yet");
     for (auto & mit : _jobs)
     {
-        Job * job = mit.second;
-        if ((*_profiles)[job->profile]->type == ProfileType::SMPI)
+        auto job = mit.second;
+        if (job->profile->type == ProfileType::SMPI)
         {
             return true;
         }
@@ -289,35 +326,36 @@ void Jobs::displayDebug() const
     XBT_DEBUG("%s", s.c_str());
 }
 
-const std::map<JobIdentifier, Job* > &Jobs::jobs() const
+const std::unordered_map<JobIdentifier, JobPtr, JobIdentifierHasher> &Jobs::jobs() const
 {
     return _jobs;
 }
 
-std::map<JobIdentifier, Job *> &Jobs::jobs()
+std::unordered_map<JobIdentifier, JobPtr, JobIdentifierHasher> &Jobs::jobs()
 {
     return _jobs;
 }
 
 int Jobs::nb_jobs() const
 {
-    return _jobs.size();
+    return static_cast<int>(_jobs.size());
 }
 
-bool job_comparator_subtime_number(const Job *a, const Job *b)
+bool job_comparator_subtime_number(const JobPtr a, const JobPtr b)
 {
     if (a->submission_time == b->submission_time)
     {
-        return *a < *b;
+        return a->id.to_string() < b->id.to_string();
     }
     return a->submission_time < b->submission_time;
 }
 
 Job::~Job()
 {
+    XBT_INFO("Job '%s' is being deleted", id.to_string().c_str());
     xbt_assert(execution_actors.size() == 0,
-               "Internal error: job %s has %d execution processes on destruction (should be 0).",
-               this->id.to_string().c_str(), (int)execution_actors.size());
+               "Internal error: job %s on destruction still has %zu execution processes (should be 0).",
+               this->id.to_string().c_str(), execution_actors.size());
 
     if (task != nullptr)
     {
@@ -336,16 +374,17 @@ bool Job::is_complete() const
     return (state == JobState::JOB_STATE_COMPLETED_SUCCESSFULLY) ||
            (state == JobState::JOB_STATE_COMPLETED_KILLED) ||
            (state == JobState::JOB_STATE_COMPLETED_FAILED) ||
+           (state == JobState::JOB_STATE_REJECTED) ||
            (state == JobState::JOB_STATE_COMPLETED_WALLTIME_REACHED);
 }
 
 // Do NOT remove namespaces in the arguments (to avoid doxygen warnings)
-Job * Job::from_json(const rapidjson::Value & json_desc,
+JobPtr Job::from_json(const rapidjson::Value & json_desc,
                      Workload * workload,
                      const std::string & error_prefix)
 {
     // Create and initialize with default values
-    Job * j = new Job;
+    auto j = std::make_shared<Job>();
     j->workload = workload;
     j->starting_time = -1;
     j->runtime = -1;
@@ -382,7 +421,7 @@ Job * Job::from_json(const rapidjson::Value & json_desc,
                error_prefix.c_str(), j->id.to_string().c_str());
     xbt_assert(json_desc["subtime"].IsNumber(), "%s: job '%s' has a non-number 'subtime' field",
                error_prefix.c_str(), j->id.to_string().c_str());
-    j->submission_time = json_desc["subtime"].GetDouble();
+    j->submission_time = static_cast<long double>(json_desc["subtime"].GetDouble());
 
     // Get walltime (optional)
     if (!json_desc.HasMember("walltime"))
@@ -393,26 +432,33 @@ Job * Job::from_json(const rapidjson::Value & json_desc,
     {
         xbt_assert(json_desc["walltime"].IsNumber(), "%s: job %s has a non-number 'walltime' field",
                    error_prefix.c_str(), j->id.to_string().c_str());
-        j->walltime = json_desc["walltime"].GetDouble();
+        j->walltime = static_cast<long double>(json_desc["walltime"].GetDouble());
     }
     xbt_assert(j->walltime == -1 || j->walltime > 0,
-               "%s: job '%s' has an invalid walltime (%g). It should either be -1 (no walltime) "
+               "%s: job '%s' has an invalid walltime (%Lg). It should either be -1 (no walltime) "
                "or a strictly positive number.",
-               error_prefix.c_str(), j->id.to_string().c_str(), (double)j->walltime);
+               error_prefix.c_str(), j->id.to_string().c_str(), j->walltime);
 
     // Get number of requested resources
     xbt_assert(json_desc.HasMember("res"), "%s: job %s has no 'res' field",
                error_prefix.c_str(), j->id.to_string().c_str());
     xbt_assert(json_desc["res"].IsInt(), "%s: job %s has a non-number 'res' field",
                error_prefix.c_str(), j->id.to_string().c_str());
-    j->requested_nb_res = json_desc["res"].GetInt();
+    xbt_assert(json_desc["res"].GetInt() >= 0, "%s: job %s has a negative 'res' field (%d)",
+               error_prefix.c_str(), j->id.to_string().c_str(), json_desc["res"].GetInt());
+    j->requested_nb_res = static_cast<unsigned int>(json_desc["res"].GetInt());
 
     // Get the job profile
     xbt_assert(json_desc.HasMember("profile"), "%s: job %s has no 'profile' field",
                error_prefix.c_str(), j->id.to_string().c_str());
     xbt_assert(json_desc["profile"].IsString(), "%s: job %s has a non-string 'profile' field",
                error_prefix.c_str(), j->id.to_string().c_str());
-    j->profile = json_desc["profile"].GetString();
+
+    // TODO raise exception when the profile does not exist.
+    std::string profile_name = json_desc["profile"].GetString();
+    xbt_assert(workload->profiles->exists(profile_name), "%s: the profile %s for job %s does not exist",
+               error_prefix.c_str(), profile_name.c_str(), j->id.to_string().c_str());
+    j->profile = workload->profiles->at(profile_name);
 
     // Let's get the JSON string which originally described the job
     // (to conserve potential fields unused by Batsim)
@@ -424,11 +470,11 @@ Job * Job::from_json(const rapidjson::Value & json_desc,
     // in the json raw description
     string json_description_tmp(buffer.GetString(), buffer.GetSize());
     /// @cond DOXYGEN_FAILS_PARSING_THIS_REGEX
-    boost::regex r(R"("id"\s*:\s*(?:"*[^(,|})]*"*)\s*)");
+    std::regex r(R"("id"\s*:\s*(?:"*[^(,|})]*"*)\s*)");
     /// @endcond
     string replacement_str = "\"id\":\"" + j->id.to_string() + "\"";
     // XBT_INFO("Before regexp: %s", json_description_tmp.c_str());
-    j->json_description = boost::regex_replace(json_description_tmp, r, replacement_str);
+    j->json_description = std::regex_replace(json_description_tmp, r, replacement_str);
 
     // Let's check that the new description is a valid JSON string
     rapidjson::Document check_doc;
@@ -479,7 +525,7 @@ Job * Job::from_json(const rapidjson::Value & json_desc,
                        "%d does not point to an integral number",
                        error_prefix.c_str(), j->id.to_string().c_str(), i);
             int host_number = mapping_array[i].GetInt();
-            xbt_assert(host_number >= 0 && (unsigned int)host_number < j->requested_nb_res,
+            xbt_assert(host_number >= 0 && static_cast<unsigned int>(host_number) < j->requested_nb_res,
                        "%s: job '%s' has a bad 'smpi_ranks_to_hosts_mapping' field: rank "
                        "%d has an invalid value %d : should be in [0,%d[",
                        error_prefix.c_str(), j->id.to_string().c_str(),
@@ -494,7 +540,7 @@ Job * Job::from_json(const rapidjson::Value & json_desc,
 }
 
 // Do NOT remove namespaces in the arguments (to avoid doxygen warnings)
-Job * Job::from_json(const std::string & json_str,
+JobPtr Job::from_json(const std::string & json_str,
                      Workload * workload,
                      const std::string & error_prefix)
 {
