@@ -42,11 +42,12 @@ struct BatsimContext
     Machines machines;                              //!< The machines
     Workloads workloads;                            //!< The workloads
     Workflows workflows;                            //!< The workflows
-    EventListMap eventListsMap;                     //!< The map of EventLists
+    std::map<std::string, EventList*> event_lists;  //!< The map of EventLists
     PajeTracer paje_tracer;                         //!< The PajeTracer
     PStateChangeTracer pstate_tracer;               //!< The PStateChangeTracer
     EnergyConsumptionTracer energy_tracer;          //!< The EnergyConsumptionTracer
     MachineStateTracer machine_state_tracer;        //!< The MachineStateTracer
+    JobsTracer jobs_tracer;                         //!< The JobsTracer
     CurrentSwitches current_switches;               //!< The current switches
 
     RedisStorage storage;                           //!< The RedisStorage
@@ -55,8 +56,9 @@ struct BatsimContext
     bool redis_enabled;                             //!< Stores whether Redis should be used
     bool submission_forward_profiles;               //!< Stores whether the profile information of submitted jobs should be sent to the scheduler
     bool registration_sched_enabled;                //!< Stores whether the scheduler will be able to register jobs and profiles during the simulation
-    bool registration_sched_finished = false;         //!< Stores whether the scheduler has finished submitting jobs.
+    bool registration_sched_finished = false;       //!< Stores whether the scheduler has finished submitting jobs.
     bool registration_sched_ack;                    //!< Stores whether Batsim will acknowledge dynamic job submission (emit JOB_SUBMITTED events)
+    bool garbage_collect_profiles = true;           //!< Stores whether Batsim will garbage collect the Profiles.
 
     bool terminate_with_last_workflow;              //!< If true, allows to ignore the jobs submitted after the last workflow termination
 
@@ -80,5 +82,7 @@ struct BatsimContext
     std::string export_prefix;                      //!< The output export prefix
     int workflow_nb_concurrent_jobs_limit;          //!< Limits the number of concurrent jobs for workflows
 
-    std::string batsim_version;                     //!< The Batsim version (got from the BATSIM_VERSION variable that is usually set by CMake)
+    std::string batsim_version;                     //!< The Batsim version (got from the BATSIM_VERSION variable that is usually set by the build system)
+
+    ~BatsimContext();
 };
