@@ -73,6 +73,10 @@ void prepare_batsim_outputs(BatsimContext * context)
             context->pstate_tracer.add_pstate_change(simgrid::s4u::Engine::get_clock(), range, pstate);
         }
     }
+
+    context->jobs_tracer.initialize(context,
+                                    context->export_prefix + "_jobs.csv",
+                                    context->export_prefix + "_schedule.csv");
 }
 
 void finalize_batsim_outputs(BatsimContext * context)
@@ -102,18 +106,15 @@ void finalize_batsim_outputs(BatsimContext * context)
         context->pstate_tracer.close_buffer();
     }
 
-    // Schedule-oriented output information
-    export_schedule_to_csv(context->export_prefix + "_schedule.csv", context);
-
-    // Job-oriented output information
-    export_jobs_to_csv(context->export_prefix + "_jobs.csv", context);
+    // Finalize both jobs and schedule output files
+    context->jobs_tracer.finalize();
 }
 
 
-WriteBuffer::WriteBuffer(const std::string & filename, int buffer_size)
+WriteBuffer::WriteBuffer(const std::string & filename, size_t buffer_size)
     : buffer_size(buffer_size)
 {
-    xbt_assert(buffer_size > 0, "Invalid buffer size (%d)", buffer_size);
+    xbt_assert(buffer_size > 0, "Invalid buffer size (%zu)", buffer_size);
     buffer = new char[buffer_size];
 
     f.open(filename, ios_base::trunc);
@@ -138,7 +139,7 @@ WriteBuffer::~WriteBuffer()
 
 void WriteBuffer::append_text(const char * text)
 {
-    const int text_length = strlen(text);
+    const size_t text_length = strlen(text);
 
     // Is the buffer big enough?
     if (buffer_pos + text_length < buffer_size)
@@ -162,14 +163,14 @@ void WriteBuffer::append_text(const char * text)
         else
         {
             // Directly write the text into the file
-            f.write(text, text_length);
+            f.write(text, static_cast<std::streamsize>(text_length));
         }
     }
 }
 
 void WriteBuffer::flush_buffer()
 {
-    f.write(buffer, buffer_pos);
+    f.write(buffer, static_cast<std::streamsize>(buffer_pos));
     buffer_pos = 0;
 }
 
@@ -217,8 +218,8 @@ void PajeTracer::initialize(const BatsimContext *context, double time)
     const int buf_size = 8 * 1024;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Coudln't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Coudln't allocate memory");
 
     // Let's write the Pajé schedule header
     // Let's write the Pajé schedule header
@@ -424,8 +425,8 @@ void PajeTracer::finalize(const BatsimContext * context, double time)
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     nb_printed = snprintf(buf, buf_size,
                           "\n"
@@ -464,18 +465,16 @@ void PajeTracer::finalize(const BatsimContext * context, double time)
     XBT_INFO("PajeTracer finalized");
 }
 
-void PajeTracer::add_job_launching(const Job * job,
-                                   const std::vector<int> & used_machine_ids,
+void PajeTracer::add_job_launching(const std::vector<int> & used_machine_ids,
                                    double time)
 {
-    (void) job;
     xbt_assert(state == INITIALIZED, "Bad addJobLaunching call: the PajeTracer object is not initialized or had been finalized");
 
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     if (_log_launchings)
     {
@@ -495,29 +494,29 @@ void PajeTracer::add_job_launching(const Job * job,
     free(buf);
 }
 
-void PajeTracer::register_new_job(const Job *job)
+void PajeTracer::register_new_job(const JobIdentifier & job_id)
 {
-    xbt_assert(_jobs.find(job) == _jobs.end(),
-               "Cannot register new job %s: it already exists", job->id.to_string().c_str());
+    xbt_assert(_jobs.find(job_id) == _jobs.end(),
+               "Cannot register new job %s: it already exists", job_id.to_cstring());
 
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     // Let's create a state value corresponding to this job
     nb_printed = snprintf(buf, buf_size,
                           "%d %s%s %s \"%s\" %s\n",
-                          DEFINE_ENTITY_VALUE, jobPrefix, job->id.to_string().c_str(),
-                          machineState, job->id.to_string().c_str(),
-                          _colors[nb_total_jobs++ % (int)_colors.size()].c_str());
+                          DEFINE_ENTITY_VALUE, jobPrefix, job_id.to_cstring(),
+                          machineState, job_id.to_cstring(),
+                          _colors[nb_total_jobs++ % _colors.size()].c_str());
     xbt_assert(nb_printed < buf_size - 1,
                "Writing error: buffer has been completely filled, some information might "
                "have been lost. Please increase Batsim's output temporary buffers' size");
     _wbuf->append_text(buf);
 
-    _jobs[job] = jobPrefix + job->id.to_string();
+    _jobs[job_id] = jobPrefix + job_id.to_string();
 
     free(buf);
 }
@@ -527,8 +526,8 @@ void PajeTracer::set_machine_idle(int machine_id, double time)
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     nb_printed = snprintf(buf, buf_size,
                           "%d %lf %s %s%d %s\n",
@@ -541,20 +540,20 @@ void PajeTracer::set_machine_idle(int machine_id, double time)
     free(buf);
 }
 
-void PajeTracer::set_machine_as_computing_job(int machine_id, const Job * job, double time)
+void PajeTracer::set_machine_as_computing_job(int machine_id, const JobIdentifier & job_id, double time)
 {
-    auto mit = _jobs.find(job);
+    auto mit = _jobs.find(job_id);
     if (mit == _jobs.end())
     {
-        register_new_job(job);
-        mit = _jobs.find(job);
+        register_new_job(job_id);
+        mit = _jobs.find(job_id);
     }
 
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     nb_printed = snprintf(buf, buf_size,
                           "%d %lf %s %s%d %s\n",
@@ -568,7 +567,7 @@ void PajeTracer::set_machine_as_computing_job(int machine_id, const Job * job, d
     free(buf);
 }
 
-void PajeTracer::add_job_kill(const Job *job, const IntervalSet & used_machine_ids,
+void PajeTracer::add_job_kill(const JobIdentifier & job_id, const IntervalSet & used_machine_ids,
                               double time, bool associate_kill_to_machines)
 {
     xbt_assert(state == INITIALIZED, "Bad addJobKill call: the PajeTracer object is not initialized or had been finalized");
@@ -576,13 +575,13 @@ void PajeTracer::add_job_kill(const Job *job, const IntervalSet & used_machine_i
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     // Let's add a kill event associated with the scheduler
     nb_printed = snprintf(buf, buf_size,
                           "%d %lf %s %s \"%s\"\n",
-                          NEW_EVENT, time, killEventKiller, killer, job->id.to_string().c_str());
+                          NEW_EVENT, time, killEventKiller, killer, job_id.to_cstring());
     xbt_assert(nb_printed < buf_size - 1,
                "Writing error: buffer has been completely filled, some information might "
                "have been lost. Please increase Batsim's output temporary buffers' size");
@@ -597,7 +596,7 @@ void PajeTracer::add_job_kill(const Job *job, const IntervalSet & used_machine_i
             nb_printed = snprintf(buf, buf_size,
                                   "%d %lf %s %s%d \"%s\"\n",
                                   NEW_EVENT, time, killEventMachine, machinePrefix, machine_id,
-                                  job->id.to_string().c_str());
+                                  job_id.to_cstring());
             xbt_assert(nb_printed < buf_size - 1,
                        "Writing error: buffer has been completely filled, some information might "
                        "have been lost. Please increase Batsim's output temporary buffers' size");
@@ -615,8 +614,8 @@ void PajeTracer::generate_colors(int color_count)
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     double hueFraction = 360.0 / color_count;
     for (int i = 0; i < color_count; ++i)
@@ -649,11 +648,11 @@ void PajeTracer::hsv_to_rgb(double h, double s, double v, double & r, double & g
     }
 
     h /= 60;            // sector 0 to 5
-    int i = floor(h);
-    float f = h-i;      // factorial part of h
-    float p = v*(1-s);
-    float q = v*(1-s*f);
-    float t = v*(1-s*(1-f));
+    int i = static_cast<int>(floor(h));
+    double f = h-i;      // factorial part of h
+    double p = v*(1-s);
+    double q = v*(1-s*f);
+    double t = v*(1-s*(1-f));
 
     switch(i)
     {
@@ -690,293 +689,13 @@ void PajeTracer::hsv_to_rgb(double h, double s, double v, double & r, double & g
     }
 }
 
-void export_jobs_to_csv(const std::string &filename, const BatsimContext *context)
-{
-    ofstream f(filename, ios_base::trunc);
-    xbt_assert(f.is_open(), "Cannot write file '%s'", filename.c_str());
-
-    // List all features (columns)
-    map<string, string> job_map;
-    vector<string> key_list = {
-        "job_id",
-        "workload_name",
-        "profile",
-        "submission_time",
-        "requested_number_of_resources",
-        "requested_time",
-        "success",
-        "final_state",
-        "starting_time",
-        "execution_time",
-        "finish_time",
-        "waiting_time",
-        "turnaround_time",
-        "stretch",
-        "allocated_resources",
-        "consumed_energy",
-        "metadata"
-    };
-
-    // Write headers (columns) to the output file
-    vector<string> row_content;
-    row_content.reserve(job_map.size());
-
-    for (string & mit : key_list)
-    {
-        row_content.push_back(mit);
-    }
-
-    f << boost::algorithm::join(row_content, ",") << "\n";
-
-    for (const auto mit : context->workloads.workloads())
-    {
-        string workload_name = mit.first;
-        const Workload * workload = mit.second;
-
-        if(workload->jobs)
-        {
-            const auto & jobs = workload->jobs->jobs();
-            for (const auto & mit : jobs)
-            {
-                Job * job = mit.second;
-
-                int success = (job->state == JobState::JOB_STATE_COMPLETED_SUCCESSFULLY);
-                bool rejected = (job->state == JobState::JOB_STATE_REJECTED);
-
-                // Update all values
-                job_map["job_id"] = job->id.job_name;
-                job_map["workload_name"] = string(workload_name);
-                job_map["profile"] = job->profile;
-                job_map["submission_time"] = to_string((double)job->submission_time);
-                job_map["requested_number_of_resources"] = to_string(job->requested_nb_res);
-                job_map["requested_time"] = to_string((double)job->walltime);
-                job_map["success"] = to_string(success);
-                job_map["final_state"] = job_state_to_string(job->state);
-                job_map["starting_time"] = rejected ? "" : to_string((double)job->starting_time);
-                job_map["execution_time"] = rejected ? "" :to_string((double)job->runtime);
-                job_map["finish_time"] = rejected ? "" :to_string((double)(job->starting_time + job->runtime));
-                job_map["waiting_time"] = rejected ? "" :to_string((double)(job->starting_time - job->submission_time));
-                job_map["turnaround_time"] = rejected ? "" :to_string((double)(job->starting_time + job->runtime - job->submission_time));
-                job_map["stretch"] = rejected ? "" :to_string((double)((job->starting_time + job->runtime - job->submission_time) / job->runtime));
-                job_map["consumed_energy"] = rejected ? "" :to_string(job->consumed_energy);
-                job_map["allocated_resources"] = job->allocation.to_string_hyphen(" ");
-                job_map["metadata"] = '"' + job->metadata + '"';
-
-
-                // Write values to the output file
-                row_content.resize(0);
-                for (string & mit : key_list)
-                {
-                    row_content.push_back(job_map[mit]);
-                }
-
-                f << boost::algorithm::join(row_content, ",") << "\n";
-            }
-        }
-    }
-
-    f.close();
-}
-
-
-void export_schedule_to_csv(const std::string &filename, const BatsimContext *context)
-{
-    ofstream f(filename, ios_base::trunc);
-    xbt_assert(f.is_open(), "Cannot write file '%s'", filename.c_str());
-
-    int nb_jobs = 0;
-    int nb_jobs_finished = 0;
-    int nb_jobs_success = 0;
-    int nb_jobs_killed = 0;
-    long double makespan = 0;
-    long double sum_waiting_time = 0;
-    long double sum_turnaround_time = 0;
-    long double sum_slowdown = 0;
-    long double max_waiting_time = 0;
-    long double max_turnaround_time = 0;
-    long double max_slowdown = 0;
-
-    map<string, string> output_map;
-
-    map<int, long double> machines_utilisation;
-    for (int i=0; i<context->machines.nb_machines(); i++)
-    {
-        machines_utilisation[i] = 0;
-    }
-
-    long double seconds_used_by_scheduler = context->microseconds_used_by_scheduler / 1e6l;
-    output_map["scheduling_time"] = to_string((double) seconds_used_by_scheduler);
-
-    // Let's compute the simulation time
-    chrono::duration<long double> diff = context->simulation_end_time - context->simulation_start_time;
-    long double seconds_used_by_the_whole_simulation = diff.count();
-    output_map["simulation_time"] = to_string((double) seconds_used_by_the_whole_simulation);
-
-    // Let's compute jobs-oriented metrics
-    for (const auto mit : context->workloads.workloads())
-    {
-        const Workload * workload = mit.second;
-
-        if(workload->jobs != nullptr)
-        {
-            const auto & jobs = workload->jobs->jobs();
-            for (const auto & mit : jobs)
-            {
-                Job * job = mit.second;
-                nb_jobs++;
-
-                if (job->is_complete())
-                {
-                    nb_jobs_finished++;
-
-                    if (job->state == JobState::JOB_STATE_COMPLETED_SUCCESSFULLY)
-                    {
-                        nb_jobs_success++;
-                    }
-                    else
-                    {
-                        nb_jobs_killed++;
-                    }
-
-
-                    long double starting_time = job->starting_time;
-                    long double waiting_time = starting_time - job->submission_time;
-                    long double completion_time = job->starting_time + job->runtime;
-                    long double turnaround_time = completion_time - job->submission_time;
-                    long double slowdown = turnaround_time / job->runtime;
-
-                    sum_waiting_time += waiting_time;
-                    sum_turnaround_time += turnaround_time;
-                    sum_slowdown += slowdown;
-
-                    if (completion_time > makespan)
-                    {
-                        makespan = completion_time;
-                    }
-
-                    if (waiting_time > max_waiting_time)
-                    {
-                        max_waiting_time = waiting_time;
-                    }
-
-                    if (turnaround_time > max_turnaround_time)
-                    {
-                        max_turnaround_time = turnaround_time;
-                    }
-
-                    if (slowdown > max_slowdown)
-                    {
-                        max_slowdown = slowdown;
-                    }
-
-                    IntervalSet & allocation = job->allocation;
-                    auto size = allocation.size();
-                    for (size_t i=0; i < size; i++)
-                    {
-                        int machine_idx = allocation[i];
-                        machines_utilisation[machine_idx] += job->runtime;
-                    }
-                }
-            }
-        }
-    }
-    double sum_time_running = 0;
-    double max_time_running = 0;
-
-    for (auto const & entry : machines_utilisation)
-    {
-        sum_time_running += (double) entry.second;
-        if ((double) entry.second > max_time_running)
-        {
-            max_time_running = (double) entry.second;
-        }
-    }
-    double mean_time_running = sum_time_running / machines_utilisation.size();
-
-    double success_rate = (double)nb_jobs_success/nb_jobs;
-
-    double mean_waiting_time = (double)sum_waiting_time/nb_jobs;
-    double mean_turnaround_time = (double)sum_turnaround_time/nb_jobs;
-    double mean_slowdown = (double)sum_slowdown/nb_jobs;
-
-    output_map["batsim_version"] = context->batsim_version;
-    output_map["nb_jobs"] = to_string(nb_jobs);
-    output_map["nb_jobs_finished"] = to_string(nb_jobs_finished);
-    output_map["nb_jobs_success"] = to_string(nb_jobs_success);
-    output_map["nb_jobs_killed"] = to_string(nb_jobs_killed);
-    output_map["nb_jobs_rejected"] = to_string(nb_jobs - nb_jobs_finished);
-    output_map["success_rate"] = to_string(success_rate);
-
-    output_map["makespan"] = to_string((double)makespan);
-    output_map["mean_waiting_time"] = to_string(mean_waiting_time);
-    output_map["mean_turnaround_time"] = to_string(mean_turnaround_time);
-    output_map["mean_slowdown"] = to_string(mean_slowdown);
-    output_map["max_waiting_time"] = to_string((double)max_waiting_time);
-    output_map["max_turnaround_time"] = to_string((double)max_turnaround_time);
-    output_map["max_slowdown"] = to_string((double)max_slowdown);
-
-    output_map["nb_computing_machines"] = to_string(context->machines.nb_machines());
-
-    XBT_INFO("jobs=%d, finished=%d, success=%d, killed=%d, success_rate=%lf",
-             nb_jobs, nb_jobs_finished, nb_jobs_success, nb_jobs_killed, success_rate);
-    XBT_INFO("makespan=%lf, scheduling_time=%lf, mean_waiting_time=%lf, mean_turnaround_time=%lf, "
-             "mean_slowdown=%lf, max_waiting_time=%lf, max_turnaround_time=%lf, max_slowdown=%lf",
-             (double)makespan, (double)seconds_used_by_scheduler,
-             (double)mean_waiting_time, (double)mean_turnaround_time, mean_slowdown,
-             (double)max_waiting_time, (double)max_turnaround_time, (double)max_slowdown);
-    XBT_INFO("mean_machines_running=%lf, max_machines_running=%lf",
-             (double)mean_time_running, (double)max_time_running);
-
-    long double total_consumed_energy = context->energy_last_job_completion - context->energy_first_job_submission;
-    output_map["consumed_joules"] = to_string((double) total_consumed_energy);
-
-    output_map["nb_machine_switches"] = to_string(context->nb_machine_switches);
-    output_map["nb_grouped_switches"] = to_string(context->nb_grouped_switches);
-
-    // Let's compute machine-related metrics
-    map<MachineState, long double> time_spent_in_each_state;
-    const vector<MachineState> machine_states = {MachineState::SLEEPING, MachineState::IDLE,
-                                                 MachineState::COMPUTING,
-                                                 MachineState::TRANSITING_FROM_SLEEPING_TO_COMPUTING,
-                                                 MachineState::TRANSITING_FROM_COMPUTING_TO_SLEEPING,
-                                                 MachineState::UNAVAILABLE};
-    for (const MachineState & state : machine_states)
-    {
-        time_spent_in_each_state[state] = 0;
-    }
-
-    for (const Machine * machine : context->machines.machines())
-    {
-        for (const MachineState & state : machine_states)
-        {
-            time_spent_in_each_state[state] += machine->time_spent_in_each_state.at(state);
-        }
-    }
-
-    for (const MachineState & state : machine_states)
-    {
-        output_map["time_" + machine_state_to_string(state)] = to_string((double)time_spent_in_each_state[state]);
-    }
-
-
-    // Let's write the output map into the file
-    vector<string> keys, values;
-    for (const auto mit : output_map)
-    {
-        keys.push_back(mit.first);
-        values.push_back(mit.second);
-    }
-
-    f << boost::algorithm::join(keys, ",") << "\n";
-    f << boost::algorithm::join(values, ",") << "\n";
-
-    f.close();
-}
+/* Part related to PStateChangeTracer */
 
 PStateChangeTracer::PStateChangeTracer()
 {
     xbt_assert(_temporary_buffer == nullptr);
-    _temporary_buffer = (char*) malloc(512 * sizeof(char));
+    _temporary_buffer = static_cast<char*>(malloc(512 * sizeof(char)));
+    xbt_assert(_temporary_buffer != NULL, "Couldn't allocate memory");
 }
 
 void PStateChangeTracer::setFilename(const string &filename)
@@ -1007,17 +726,17 @@ void PStateChangeTracer::add_pstate_change(double time, const IntervalSet & mach
     xbt_assert(_wbuf != nullptr);
 
     const string machines_as_string = machines.to_string_hyphen(" ", "-");
-    const int minimum_buf_size = 256 + machines_as_string.size();
+    const size_t minimum_buf_size = 256 + machines_as_string.size();
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
 
     // Increases the buffer size if needed
-    _temporary_buffer = (char*) realloc(_temporary_buffer, minimum_buf_size);
-    xbt_assert(_temporary_buffer != 0, "Couldn't allocate memory!");
+    _temporary_buffer = static_cast<char*>(realloc(_temporary_buffer, minimum_buf_size));
+    xbt_assert(_temporary_buffer != NULL, "Couldn't allocate memory!");
 
     nb_printed = snprintf(_temporary_buffer, minimum_buf_size, "%g,%s,%d\n",
                           time, machines_as_string.c_str(), pstate_after);
-    xbt_assert(nb_printed < minimum_buf_size - 1,
+    xbt_assert(nb_printed < static_cast<int>(minimum_buf_size) - 1,
                "Writing error: buffer has been completely filled, some information might "
                "have been lost. Please increase Batsim's output temporary buffers' size");
     _wbuf->append_text(_temporary_buffer);
@@ -1039,6 +758,7 @@ void PStateChangeTracer::close_buffer()
 }
 
 
+/* Part related to EnergyConsumptionTracer */
 
 EnergyConsumptionTracer::~EnergyConsumptionTracer()
 {
@@ -1104,7 +824,7 @@ long double EnergyConsumptionTracer::add_entry(double date, char event_type)
     long double energy = _context->machines.total_consumed_energy(_context);
     long double wattmin = _context->machines.total_wattmin(_context);
 
-    long double time_diff = (long double)date - _last_entry_date;
+    long double time_diff = static_cast<long double>(date) - _last_entry_date;
     long double energy_diff = energy - _last_entry_energy;
     long double epower = -1;
 
@@ -1116,13 +836,13 @@ long double EnergyConsumptionTracer::add_entry(double date, char event_type)
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     if (epower != -1)
     {
         nb_printed = snprintf(buf, buf_size, "%g,%Lg,%c,%Lg,%g\n",
-                              date, energy, event_type, wattmin, (double)epower);
+                              date, energy, event_type, wattmin, static_cast<double>(epower));
     }
     else
     {
@@ -1137,12 +857,15 @@ long double EnergyConsumptionTracer::add_entry(double date, char event_type)
 
     free(buf);
 
-    _last_entry_date = date;
+    _last_entry_date = static_cast<long double>(date);
     _last_entry_energy = energy;
 
     return energy;
 }
 
+
+
+/* Part related to MachineStateTracer */
 
 MachineStateTracer::~MachineStateTracer()
 {
@@ -1190,8 +913,8 @@ void MachineStateTracer::write_machine_states(double date)
     const int buf_size = 256;
     int nb_printed;
     (void) nb_printed; // Avoids a warning if assertions are ignored
-    char * buf = (char*) malloc(sizeof(char) * buf_size);
-    xbt_assert(buf != 0, "Couldn't allocate memory");
+    char * buf = static_cast<char*>(malloc(sizeof(char) * buf_size));
+    xbt_assert(buf != NULL, "Couldn't allocate memory");
 
     nb_printed = snprintf(buf, buf_size, "%g,%d,%d,%d,%d,%d\n",
                           date,
@@ -1220,6 +943,276 @@ void MachineStateTracer::close_buffer()
 {
     xbt_assert(_wbuf != nullptr);
 
+    delete _wbuf;
+    _wbuf = nullptr;
+}
+
+/* Part related to JobsTracer */
+
+JobsTracer::~JobsTracer()
+{
+    if (_wbuf != nullptr)
+    {
+        delete _wbuf;
+        _wbuf = nullptr;
+    }
+}
+
+void JobsTracer::initialize(BatsimContext *context,
+                       const string & jobs_filename,
+                       const string & schedule_filename)
+{
+    xbt_assert(_wbuf == nullptr, "Double call of JobsTracer::initialize");
+    _wbuf = new WriteBuffer(jobs_filename);
+    _context = context;
+    _schedule_filename = schedule_filename;
+
+    // Prepare for jobs output file
+    _job_keys = {
+        "job_id",
+        "workload_name",
+        "profile",
+        "submission_time",
+        "requested_number_of_resources",
+        "requested_time",
+        "success",
+        "final_state",
+        "starting_time",
+        "execution_time",
+        "finish_time",
+        "waiting_time",
+        "turnaround_time",
+        "stretch",
+        "allocated_resources",
+        "consumed_energy",
+        "metadata"
+    };
+    string header = boost::algorithm::join(_job_keys, ",") + "\n";
+    _row_content.reserve(_job_keys.size());
+    _wbuf->append_text(header.c_str());
+    _wbuf->flush_buffer();
+
+    // Prepare for schedule output file
+    for (int i = 0; i < static_cast<int>(context->machines.nb_machines()); ++i)
+    {
+        _machines_utilization[i] = 0;
+    }
+}
+
+void JobsTracer::finalize()
+{
+    // Finalize jobs output file
+    flush();
+    close_buffer();
+
+    // Write the schedule output file
+    ofstream f(_schedule_filename, ios_base::trunc);
+    xbt_assert(f.is_open(), "Cannot write file '%s'", _schedule_filename.c_str());
+    map<string, string> output_map;
+
+    long double seconds_used_by_scheduler = _context->microseconds_used_by_scheduler / 1e6l;
+    output_map["scheduling_time"] = to_string(static_cast<double>(seconds_used_by_scheduler));
+
+    // Let's compute the simulation time
+    chrono::duration<long double> diff = _context->simulation_end_time - _context->simulation_start_time;
+    long double seconds_used_by_the_whole_simulation = diff.count();
+    output_map["simulation_time"] = to_string(static_cast<double>(seconds_used_by_the_whole_simulation));
+
+    double sum_time_running = 0;
+    double max_time_running = 0;
+    for (auto const & entry : _machines_utilization)
+    {
+        sum_time_running += static_cast<double>(entry.second);
+        if (static_cast<double>(entry.second) > max_time_running)
+        {
+            max_time_running = static_cast<double>(entry.second);
+        }
+    }
+    double mean_time_running = sum_time_running / _machines_utilization.size();
+
+    double success_rate = static_cast<double>(_nb_jobs_success)/_nb_jobs;
+
+    double mean_waiting_time = static_cast<double>(_sum_waiting_time)/_nb_jobs;
+    double mean_turnaround_time = static_cast<double>(_sum_turnaround_time)/_nb_jobs;
+    double mean_slowdown = static_cast<double>(_sum_slowdown)/_nb_jobs;
+
+    output_map["batsim_version"] = _context->batsim_version;
+    output_map["nb_jobs"] = to_string(_nb_jobs);
+    output_map["nb_jobs_finished"] = to_string(_nb_jobs_finished);
+    output_map["nb_jobs_success"] = to_string(_nb_jobs_success);
+    output_map["nb_jobs_killed"] = to_string(_nb_jobs_killed);
+    output_map["nb_jobs_rejected"] = to_string(_nb_jobs_rejected);
+    output_map["success_rate"] = to_string(success_rate);
+
+    output_map["makespan"] = to_string(static_cast<double>(_makespan));
+    output_map["mean_waiting_time"] = to_string(mean_waiting_time);
+    output_map["mean_turnaround_time"] = to_string(mean_turnaround_time);
+    output_map["mean_slowdown"] = to_string(mean_slowdown);
+    output_map["max_waiting_time"] = to_string(static_cast<double>(_max_waiting_time));
+    output_map["max_turnaround_time"] = to_string(static_cast<double>(_max_turnaround_time));
+    output_map["max_slowdown"] = to_string(static_cast<double>(_max_slowdown));
+
+    output_map["nb_computing_machines"] = to_string(_context->machines.nb_machines());
+
+    XBT_INFO("jobs=%d, finished=%d, success=%d, killed=%d, success_rate=%lf",
+             _nb_jobs, _nb_jobs_finished, _nb_jobs_success, _nb_jobs_killed, success_rate);
+    XBT_INFO("makespan=%lf, scheduling_time=%lf, mean_waiting_time=%lf, mean_turnaround_time=%lf, "
+             "mean_slowdown=%lf, max_waiting_time=%lf, max_turnaround_time=%lf, max_slowdown=%lf",
+             static_cast<double>(_makespan), static_cast<double>(seconds_used_by_scheduler),
+             static_cast<double>(mean_waiting_time), static_cast<double>(mean_turnaround_time), mean_slowdown,
+             static_cast<double>(_max_waiting_time), static_cast<double>(_max_turnaround_time), static_cast<double>(_max_slowdown));
+    XBT_INFO("mean_machines_running=%lf, max_machines_running=%lf",
+             static_cast<double>(mean_time_running), static_cast<double>(max_time_running));
+
+    long double total_consumed_energy = _context->energy_last_job_completion - _context->energy_first_job_submission;
+    output_map["consumed_joules"] = to_string(static_cast<double>(total_consumed_energy));
+
+    output_map["nb_machine_switches"] = to_string(_context->nb_machine_switches);
+    output_map["nb_grouped_switches"] = to_string(_context->nb_grouped_switches);
+
+    // Let's compute machine-related metrics
+    map<MachineState, long double> time_spent_in_each_state;
+    const vector<MachineState> machine_states = {MachineState::SLEEPING, MachineState::IDLE,
+                                                 MachineState::COMPUTING,
+                                                 MachineState::TRANSITING_FROM_SLEEPING_TO_COMPUTING,
+                                                 MachineState::TRANSITING_FROM_COMPUTING_TO_SLEEPING,
+                                                 MachineState::UNAVAILABLE};
+    for (const MachineState & state : machine_states)
+    {
+        time_spent_in_each_state[state] = 0;
+    }
+
+    for (const Machine * machine : _context->machines.machines())
+    {
+        for (const MachineState & state : machine_states)
+        {
+            time_spent_in_each_state[state] += machine->time_spent_in_each_state.at(state);
+        }
+    }
+
+    for (const MachineState & state : machine_states)
+    {
+        output_map["time_" + machine_state_to_string(state)] = to_string(static_cast<double>(time_spent_in_each_state[state]));
+    }
+
+    // Let's write the output map into the file
+    vector<string> keys, values;
+    for (const auto & mit : output_map)
+    {
+        keys.push_back(mit.first);
+        values.push_back(mit.second);
+    }
+
+    f << boost::algorithm::join(keys, ",") << "\n";
+    f << boost::algorithm::join(values, ",") << "\n";
+
+    f.close();
+}
+
+void JobsTracer::write_job(const JobPtr job)
+{
+    int success = (job->state == JobState::JOB_STATE_COMPLETED_SUCCESSFULLY);
+    bool rejected = (job->state == JobState::JOB_STATE_REJECTED);
+
+    // Update counters for the schedule metrics
+    _nb_jobs++;
+    if (job->is_complete())
+    {
+        if (rejected)
+        {
+            _nb_jobs_rejected++;
+        }
+        else
+        {
+            _nb_jobs_finished++;
+
+            if (success)
+            {
+                _nb_jobs_success++;
+            }
+            else
+            {
+                _nb_jobs_killed++;
+            }
+
+            long double starting_time = job->starting_time;
+            long double waiting_time = starting_time - job->submission_time;
+            long double completion_time = job->starting_time + job->runtime;
+            long double turnaround_time = completion_time - job->submission_time;
+            long double slowdown = turnaround_time / job->runtime;
+
+            _sum_waiting_time += waiting_time;
+            _sum_turnaround_time += turnaround_time;
+            _sum_slowdown += slowdown;
+
+            if (completion_time > _makespan)
+            {
+                _makespan = completion_time;
+            }
+
+            if (waiting_time > _max_waiting_time)
+            {
+                _max_waiting_time = waiting_time;
+            }
+
+            if (turnaround_time > _max_turnaround_time)
+            {
+                _max_turnaround_time = turnaround_time;
+            }
+
+            if (slowdown > _max_slowdown)
+            {
+                _max_slowdown = slowdown;
+            }
+
+            const IntervalSet & allocation = job->allocation;
+            for (size_t i = 0; i < allocation.size(); ++i)
+            {
+                _machines_utilization[allocation[static_cast<int>(i)]] += job->runtime;
+            }
+        }
+    }
+    else
+    {
+        xbt_die("Job %s did not complete", job->id.job_name().c_str());
+    }
+
+    // Set all values to be written
+    _job_map["job_id"] = job->id.job_name();
+    _job_map["workload_name"] = job->workload->name;
+    _job_map["profile"] = (job->profile)->name;
+    _job_map["submission_time"] = to_string(static_cast<double>(job->submission_time));
+    _job_map["requested_number_of_resources"] = to_string(job->requested_nb_res);
+    _job_map["requested_time"] = to_string(static_cast<double>(job->walltime));
+    _job_map["success"] = to_string(success);
+    _job_map["final_state"] = job_state_to_string(job->state);
+    _job_map["starting_time"] = rejected ? "" : to_string(static_cast<double>(job->starting_time));
+    _job_map["execution_time"] = rejected ? "" : to_string(static_cast<double>(job->runtime));
+    _job_map["finish_time"] = rejected ? "" : to_string(static_cast<double>(job->starting_time + job->runtime));
+    _job_map["waiting_time"] = rejected ? "" : to_string(static_cast<double>(job->starting_time - job->submission_time));
+    _job_map["turnaround_time"] = rejected ? "" : to_string(static_cast<double>(job->starting_time + job->runtime - job->submission_time));
+    _job_map["stretch"] = rejected ? "" : to_string(static_cast<double>((job->starting_time + job->runtime - job->submission_time) / job->runtime));
+    _job_map["consumed_energy"] = rejected ? "" : to_string(job->consumed_energy);
+    _job_map["allocated_resources"] = job->allocation.to_string_hyphen(" ");
+    _job_map["metadata"] = '"' + job->metadata + '"';
+    // And then write them
+    for (string & mit : _job_keys)
+    {
+        _row_content.push_back(_job_map[mit]);
+    }
+    _wbuf->append_text((boost::algorithm::join(_row_content, ",") + "\n").c_str());
+    _row_content.clear();
+}
+
+void JobsTracer::flush()
+{
+    xbt_assert(_wbuf != nullptr);
+    _wbuf->flush_buffer();
+}
+
+void JobsTracer::close_buffer()
+{
+    xbt_assert(_wbuf != nullptr);
     delete _wbuf;
     _wbuf = nullptr;
 }
