@@ -5,16 +5,18 @@
 
 #pragma once
 
-#include <map>
+#include <unordered_map>
 #include <vector>
 #include <deque>
+#include <memory>
 
 #include <rapidjson/document.h>
 
-#include <simgrid/msg.h>
 #include <simgrid/s4u.hpp>
 
 #include <intervalset.hpp>
+
+#include "pointers.hpp"
 
 class Profiles;
 struct Profile;
@@ -24,8 +26,9 @@ struct Job;
 /**
  * @brief A simple structure used to identify one job
  */
-struct JobIdentifier
+class JobIdentifier
 {
+public:
     /**
      * @brief Creates an empty JobIdentifier
      */
@@ -53,6 +56,12 @@ struct JobIdentifier
     std::string to_string() const;
 
     /**
+     * @brief Returns a null-terminated C string of the JobIdentifier representation.
+     * @return A null-terminated C string of the JobIdentifier representation.
+     */
+    const char * to_cstring() const;
+
+    /**
      * @brief Returns whether the fields are lexically valid.
      * @details None of the fields should contain a '!'.
      * @param[out] reason Empty if valid.
@@ -66,9 +75,29 @@ struct JobIdentifier
      */
     void check_lexically_valid() const;
 
-public:
-    std::string workload_name; //!< The name of the workload the job belongs to
-    std::string job_name; //!< The job unique name inside its workload
+    /**
+     * @brief Returns the workload name.
+     * @return The workload name.
+     */
+    std::string workload_name() const;
+
+    /**
+     * @brief Returns the job name within the workload.
+     * @return The job name within the workload.
+     */
+    std::string job_name() const;
+
+private:
+    /**
+     * @brief Computes the string representation of the JobIdentifier.
+     * @return The string representation of the JobIdentifier.
+     */
+    std::string representation() const;
+
+private:
+    std::string _workload_name; //!< The name of the workload the job belongs to
+    std::string _job_name; //!< The job unique name inside its workload
+    std::string _representation; //!< Stores a string representation of the JobIdentifier
 };
 
 /**
@@ -79,6 +108,24 @@ public:
  */
 bool operator<(const JobIdentifier & ji1, const JobIdentifier & ji2);
 
+/**
+ * @brief Compares two JobIdentifier thanks to their string representations
+ * @param[in] ji1 The first JobIdentifier
+ * @param[in] ji2 The second JobIdentifier
+ * @return ji1.to_string() == ji2.to_string()
+ */
+bool operator==(const JobIdentifier & ji1, const JobIdentifier & ji2);
+
+//! Functor to hash a JobIdentifier
+struct JobIdentifierHasher
+{
+    /**
+     * @brief Hashes a JobIdentifier.
+     * @param[in] id The JobIdentifier to hash.
+     * @return Whatever is returned by std::hash to match C++ conventions.
+     */
+    std::size_t operator()(const JobIdentifier & id) const;
+};
 
 /**
  * @brief Contains the different states a job can be in
@@ -107,7 +154,7 @@ struct BatTask
      * @param[in] parent_job The job that owns the task
      * @param[in] profile The profile that corresponds to the task
      */
-    BatTask(Job * parent_job, Profile * profile);
+    BatTask(JobPtr parent_job, ProfilePtr profile);
 
     /**
      * @brief Battask cannot be copied.
@@ -135,12 +182,12 @@ private:
     void compute_leaf_progress();
 
 public:
-    Job * parent_job; //!< The parent job that owns this task
-    Profile * profile; //!< The task profile. The corresponding profile tells how the job should be computed
-    Profile * io_profile = nullptr; //!< The task additional io profile. This profile, if defined will b merge to the task profile before execution
+    JobPtrWeak parent_job; //!< The parent job that owns this task
+    ProfilePtr profile; //!< The task profile. The corresponding profile tells how the job should be computed
+    ProfilePtr io_profile = nullptr; //!< The task additional io profile. This profile, if defined will b merge to the task profile before execution
 
-    // Manage MSG profiles
-    msg_task_t ptask = nullptr; //!< The final task to execute (only set for BatTask leaves with MSG profiles)
+    // Manage parallel profiles
+    simgrid::s4u::ExecPtr ptask = nullptr; //!< The final task to execute (only set for BatTask leaves with parallel profiles)
 
     // manage Delay profile
     double delay_task_start = -1; //!< Stores when the task started its execution, in order to compute its progress afterwards (only set for BatTask leaves with delay profiles)
@@ -148,7 +195,7 @@ public:
 
     // manage sequential profile
     std::vector<BatTask*> sub_tasks; //!< List of sub tasks that must be executed sequentially. Only set for BatTask non-leaves with sequential profiles at the moment, but it may be used for parallel composition in the future.
-    unsigned int current_task_index = -1; //!< Index of the task that is currently being executed in the sub_tasks vector. Only set for BatTask non-leaves with sequential profiles.
+    unsigned int current_task_index = static_cast<unsigned int>(-1); //!< Index of the task that is currently being executed in the sub_tasks vector. Only set for BatTask non-leaves with sequential profiles.
     double current_task_progress_ratio = 0; //!< Gives the progress of the current task from 0 to 1. Only set for BatTask non-leaves with sequential profiles.
 };
 
@@ -186,7 +233,7 @@ struct Job
     long double consumed_energy; //!< The sum, for all machine on which the job has been allocated, of the consumed energy (in Joules) during the job execution time (consumed_energy_after_job_completion - consumed_energy_before_job_start)
 
     // User inputs
-    std::string profile; //!< The job profile name. The corresponding profile tells how the job should be computed
+    ProfilePtr profile; //!< A pointer to the job profile. The profile tells how the job should be computed
     long double submission_time; //!< The job submission time: The time at which the becomes available
     long double walltime = -1; //!< The job walltime: if the job is executed for more than this amount of time, it will be killed. Set at -1 to disable this behavior
     unsigned int requested_nb_res; //!< The number of resources the job is requested to be executed on
@@ -207,7 +254,7 @@ public:
      * @return The newly allocated Job
      * @pre The JSON description of the job is valid
      */
-    static Job * from_json(const rapidjson::Value & json_desc,
+    static JobPtr from_json(const rapidjson::Value & json_desc,
                            Workload * workload,
                            const std::string & error_prefix = "Invalid JSON job");
 
@@ -219,7 +266,7 @@ public:
      * @return The newly allocated Job
      * @pre The JSON description of the job is valid
      */
-    static Job * from_json(const std::string & json_str,
+    static JobPtr from_json(const std::string & json_str,
                            Workload * workload,
                            const std::string & error_prefix = "Invalid JSON job");
     /**
@@ -243,7 +290,7 @@ bool operator<(const Job & j1, const Job & j2);
  * @param[in] b The second job
  * @return True if and only if the first job's submission time is lower than the second job's submission time
  */
-bool job_comparator_subtime_number(const Job * a, const Job * b);
+bool job_comparator_subtime_number(const JobPtr a, const JobPtr b);
 
 /**
  * @brief Stores all the jobs of a workload
@@ -292,21 +339,21 @@ public:
      * @param[in] job_id The job id
      * @return A pointer to the job associated to the given job id
      */
-    Job * operator[](JobIdentifier job_id);
+    JobPtr operator[](JobIdentifier job_id);
 
     /**
      * @brief Accesses one job thanks to its unique name (const version)
      * @param[in] job_id The job id
      * @return A (const) pointer to the job associated to the given job id
      */
-    const Job * operator[](JobIdentifier job_id) const;
+    const JobPtr operator[](JobIdentifier job_id) const;
 
     /**
      * @brief Accesses one job thanks to its unique id
      * @param[in] job_id The job unique id
      * @return A pointer to the job associated to the given job id
      */
-    Job * at(JobIdentifier job_id);
+    JobPtr at(JobIdentifier job_id);
 
     /**
      * @brief Accesses one job thanks to its unique id (const version)
@@ -314,21 +361,29 @@ public:
      * @return A (const) pointer to the job associated to the given job
      * name
      */
-    const Job * at(JobIdentifier job_id) const;
+    const JobPtr at(JobIdentifier job_id) const;
 
     /**
      * @brief Adds a job into a Jobs instance
      * @param[in] job The job to add
      * @pre No job with the same name exist in the Jobs instance
      */
-    void add_job(Job * job);
+    void add_job(JobPtr job);
+
+    /**
+     * @brief Deletes a job
+     * @param[in] job_id The identifier of the job to delete
+     * @param[in] garbage_collect_profiles Whether to garbage collect its profiles
+     */
+    void delete_job(const JobIdentifier & job_id,
+                    const bool & garbage_collect_profiles);
 
     /**
      * @brief Allows to know whether a job exists
      * @param[in] job_id The unique job name
      * @return True if and only if a job with the given job name exists
      */
-    bool exists(JobIdentifier job_id) const;
+    bool exists(const JobIdentifier & job_id) const;
 
     /**
      * @brief Allows to know whether the Jobs contains any SMPI job
@@ -342,16 +397,16 @@ public:
     void displayDebug() const;
 
     /**
-     * @brief Returns a copy of the std::map which contains the jobs
-     * @return A copy of the std::map which contains the jobs
+     * @brief Returns a copy of the map that contains the jobs
+     * @return A copy of the map that contains the jobs
      */
-    const std::map<JobIdentifier, Job*> & jobs() const;
+    const std::unordered_map<JobIdentifier, JobPtr, JobIdentifierHasher> & jobs() const;
 
     /**
-     * @brief Returns a reference to the std::map which contains the jobs
-     * @return A reference to the std::map which contains the jobs
+     * @brief Returns a reference to the map that contains the jobs
+     * @return A reference to the map that contains the jobs
      */
-    std::map<JobIdentifier, Job*> & jobs();
+    std::unordered_map<JobIdentifier, JobPtr, JobIdentifierHasher> & jobs();
 
     /**
      * @brief Returns the number of jobs of the Jobs instance
@@ -360,7 +415,8 @@ public:
     int nb_jobs() const;
 
 private:
-    std::map<JobIdentifier, Job*> _jobs; //!< The std::map which contains the jobs
+    std::unordered_map<JobIdentifier, JobPtr, JobIdentifierHasher> _jobs; //!< The map that contains the jobs
+    std::unordered_map<JobIdentifier, bool, JobIdentifierHasher> _jobs_met; //!< Stores the jobs id already met during the simulation
     Profiles * _profiles = nullptr; //!< The profiles associated with the jobs
     Workload * _workload = nullptr; //!< The Workload the jobs belong to
 };
