@@ -32,9 +32,9 @@ struct ServerData
      */
     struct SubmitterCounters
     {
-        int expected_nb_submitters = -1;    //!< The expected number of submitters
-        int nb_submitters = 0;              //!< The number of submitters
-        int nb_submitters_finished = 0;     //!< The number of finished submitters
+        unsigned int expected_nb_submitters = static_cast<unsigned int>(-1); //!< The expected number of submitters
+        unsigned int nb_submitters = 0; //!< The number of submitters
+        unsigned int nb_submitters_finished = 0; //!< The number of finished submitters
     };
 
     BatsimContext * context = nullptr; //!< The BatsimContext
@@ -54,6 +54,7 @@ struct ServerData
     std::map<std::string, Submitter*> submitters;   //!< The submitters
     std::unordered_map<SubmitterType, SubmitterCounters> submitter_counters; //!< A map of counters for Job, Event and Workflow Submitters
     std::map<JobIdentifier, Submitter*> origin_of_jobs; //!< Stores whether a Submitter must be notified on job completion
+    std::vector<JobIdentifier> jobs_to_be_deleted; //!< Stores the job_ids to be deleted after sending a message
 };
 
 /**
