@@ -5,16 +5,14 @@
 #include "jobs.hpp"
 
 /**
- * @brief Execute tasks that correspond to parallel task (MSG) profiles
+ * @brief Execute tasks that correspond to parallel task profiles
  * @param[in,out] btask The task to execute. Progress information is stored within it.
  * @param[in] allocation The hosts where the task can be executed
  * @param[in,out] remaining_time The remaining time of the current task. It will be automatically killed if 0 is reached.
  * @param[in] context The BatsimContext
- * @param[in,out] cleanup_data Used to clean data on process kill (happens when the scheduler requests jobs to be killed).
- * @return The profile return code on success, -1 on timeout (remaining time reached 0)
+ * @return The profile return code on success, -1 on timeout (remaining time reached 0), -2 on task cancelled (issued by another SimGrid actor)
  */
-int execute_msg_task(BatTask * btask,
+int execute_parallel_task(BatTask * btask,
                      const SchedulingAllocation* allocation,
                      double * remaining_time,
-                     BatsimContext * context,
-                     CleanExecuteTaskData * cleanup_data);
+                     BatsimContext * context);
