@@ -4,10 +4,12 @@ import pytest
 import subprocess
 from collections import namedtuple
 from os.path import abspath, basename, dirname, realpath
+import helper
 
 Workload = namedtuple('Workload', ['name', 'filename'])
 Platform = namedtuple('Platform', ['name', 'filename'])
-Algorithm = namedtuple('Platform', ['name', 'sched_implem', 'sched_algo_name'])
+ExternalEvents = namedtuple('ExternalEvents', ['name', 'filename'])
+Algorithm = namedtuple('Algorithm', ['name', 'sched_implem', 'sched_algo_name'])
 
 RedisMode = namedtuple('RedisMode', ['name', 'enabled'])
 
@@ -23,15 +25,30 @@ def generate_workloads(workload_dir, definition, accepted_names):
         filename=abspath(f'{workload_dir}/{filename}'))
         for name, filename in definition.items() if name in accepted_names]
 
+def generate_external_events(external_events_dir, definition, accepted_names):
+    return [ExternalEvents(
+        name=name,
+        filename=abspath(f'{external_events_dir}/{filename}'))
+        for name, filename in definition.items() if name in accepted_names]
+
 def generate_batsched_algorithms(definition, accepted_names):
     return [Algorithm(name=name, sched_implem='batsched', sched_algo_name=sched_algo_name)
         for name, sched_algo_name in definition.items() if name in accepted_names]
+
+def generate_pybatsim_algorithms(definition, accepted_names):
+    return [Algorithm(name=name, sched_implem='pybatsim', sched_algo_name=sched_algo_name)
+        for name, sched_algo_name in definition.items() if name in accepted_names]
+
+def pytest_addoption(parser):
+    parser.addoption("--with-valgrind", action="store_true", default=False, help="runs batsim under valgrind memcheck analysis")
 
 def pytest_generate_tests(metafunc):
     script_dir = dirname(realpath(__file__))
     batsim_dir = realpath(f'{script_dir}/..')
     platform_dir = realpath(f'{batsim_dir}/platforms')
     workload_dir = realpath(f'{batsim_dir}/workloads')
+    external_events_dir = realpath(f'{batsim_dir}/events')
+    helper.set_with_valgrind(metafunc.config.option.with_valgrind)
 
     #############################
     # Define the various inputs #
@@ -42,6 +59,7 @@ def pytest_generate_tests(metafunc):
         "cluster512": "cluster512.xml",
         "energy128notopo": "energy_platform_homogeneous_no_net_128.xml",
         "energy128cluster": "cluster_energy_128.xml",
+        "properties_platform": "properties_example.xml",
     }
     energy_platforms = ["energy128notopo", "energy128cluster"]
 
@@ -56,6 +74,7 @@ def pytest_generate_tests(metafunc):
         "compute1": "test_one_computation_job.json",
         "computetot1": "test_one_computation_job_tot.json",
         "farfuture": "test_long_workload.json",
+        "genome": "GENOME.d.351024866.5.dax",
         "long": "test_batsim_paper_workload_seed1.json",
         "mixed": "test_various_profile_types.json",
         "samesubmittime": "test_same_submit_time.json",
@@ -74,6 +93,7 @@ def pytest_generate_tests(metafunc):
     dynsub_workloads = ["delay1", "mixed"]
     moldable_perf_degradation_workloads = ['compute1', 'computetot1']
     energymini_workloads = ['energymini0', 'energymini50', 'energymini100']
+    workflows = ['genome']
 
     # Algorithms
     algorithms_def = {
@@ -90,10 +110,18 @@ def pytest_generate_tests(metafunc):
         "sequencer": "sequencer",
         "sleeper": "sleeper",
         "submitter": "submitter",
+        "py_filler": "fillerSched",
+        "py_filler_events": "fillerSchedWithEvents",
     }
     basic_algorithms = ["fcfs", "easyfast", "filler"]
-    energy_algorithms = ["sleeper", "energywatcher"]
+    energy_algorithms = ["sleeper"] #fixme: enable "energywatcher" once algo fixed
     metadata_algorithms = ['filler', 'submitter']
+
+    # External Events
+    external_events_def = {
+        "simple": "test_events_4hosts.txt",
+        "generic": "test_generic_event.txt",
+    }
 
     ##########################################
     # Generate fixtures from the definitions #
@@ -107,10 +135,14 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize('small_platform', generate_platforms(platform_dir, platforms_def, ['small']))
     if 'cluster_platform' in metafunc.fixturenames:
         metafunc.parametrize('cluster_platform', generate_platforms(platform_dir, platforms_def, ['cluster512']))
+    if 'properties_platform' in metafunc.fixturenames:
+        metafunc.parametrize('properties_platform', generate_platforms(platform_dir, platforms_def, ['properties_platform']))
 
     # Workloads
     if 'workload' in metafunc.fixturenames:
         metafunc.parametrize('workload', generate_workloads(workload_dir, workloads_def, [key for key in workload_def]))
+    if 'workflow' in metafunc.fixturenames:
+        metafunc.parametrize('workflow', generate_workloads(workload_dir, workloads_def, [key for key in workflows]))
     if 'one_job_workload' in metafunc.fixturenames:
         metafunc.parametrize('one_job_workload', generate_workloads(workload_dir, workloads_def, one_job_workloads))
     if 'small_workload' in metafunc.fixturenames:
@@ -135,6 +167,14 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize('long_workload', generate_workloads(workload_dir, workloads_def, ['long']))
     if 'delaysequences_workload' in metafunc.fixturenames:
         metafunc.parametrize('delaysequences_workload', generate_workloads(workload_dir, workloads_def, ['delaysequences']))
+    if 'mixed_workload' in metafunc.fixturenames:
+        metafunc.parametrize('mixed_workload', generate_workloads(workload_dir, workloads_def, ['mixed']))
+
+    # External Events
+    if 'simple_events' in metafunc.fixturenames:
+        metafunc.parametrize('simple_events', generate_external_events(external_events_dir, external_events_def, ['simple']))
+    if 'generic_events' in metafunc.fixturenames:
+        metafunc.parametrize('generic_events', generate_external_events(external_events_dir, external_events_def, ['generic']))
 
     # Algorithms
     if 'basic_algorithm' in metafunc.fixturenames:
@@ -157,6 +197,11 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize('random_algorithm', generate_batsched_algorithms(algorithms_def, ['random']))
     if 'idle_sleeper_algorithm' in metafunc.fixturenames:
         metafunc.parametrize('idle_sleeper_algorithm', generate_batsched_algorithms(algorithms_def, ['idlesleeper']))
+
+    if 'pybatsim_filler_algorithm' in metafunc.fixturenames:
+        metafunc.parametrize('pybatsim_filler_algorithm', generate_pybatsim_algorithms(algorithms_def, ['py_filler']))
+    if 'pybatsim_filler_events_algorithm' in metafunc.fixturenames:
+        metafunc.parametrize('pybatsim_filler_events_algorithm', generate_pybatsim_algorithms(algorithms_def, ['py_filler_events']))
 
     # Misc. fixtures.
     if 'redis_mode' in metafunc.fixturenames:
